Enable cppcoreguidelines-narrowing-conversions
Make the previously-implicit narrowing conversions explicit with static_cast
(behavior-preserving: the implicit conversion already occurred; the cast just
makes it explicit to the same type). Sites span defs/{tensor,math,nn,schema,
parser}, inliner, shape inference, the IR and pb converter, and two
version-converter adapters.

Two narrowings are fixed at the root instead of with a cast:
- einsum shape inference: a label counter and its map value type were size_t
  but only hold a small int dim index, so they are retyped to int.
- MathOpTwoIntegers now takes and returns int64_t rather than int. It operates
  on dim values (int64_t), so the int signature both forced casts and truncated
  dims above INT_MAX.

Enables the check via both its bugprone- and cppcoreguidelines- alias names.

+++ onnx/common/ir_pb_converter.cc
@@ -424,7 +424,7 @@ std::unique_ptr<Graph> ImportModelProto(const ModelProto& mp) {
     return nullptr;
   }
 
-  std::unique_ptr<Graph> g(graphProtoToGraph(mp.graph(), false, mp.ir_version()));
+  std::unique_ptr<Graph> g(graphProtoToGraph(mp.graph(), false, static_cast<int>(mp.ir_version())));
   for (int i = 0; i < mp.opset_import_size(); i++) {
     OpSetID new_opset_version(mp.opset_import(i).domain(), mp.opset_import(i).version());
     g->forSelfAndEachSubGraph(
