+++ onnx/defs/parser.cc
@@ -620,7 +620,7 @@ Common::Status OnnxParser::Parse(TensorProto& tensorProto, const TypeProto& tens
               return ParseError("Mismatch between data type and value: %d, %d", elem_type, intval);
             }
             // NOLINTNEXTLINE(bugprone-narrowing-conversions)
-            tensorProto.add_int32_data(intval);
+            tensorProto.add_int32_data(static_cast<int32_t>(intval));
             break;
           case TensorProto::DataType::TensorProto_DataType_INT64:
             PARSE_TOKEN(intval);
