+++ onnx/version_converter/adapters/split_17_18.h
@@ -20,7 +20,7 @@ class Split_17_18 : public Adapter {
 
   void adapt_split_17_18(const std::shared_ptr<Graph>& /*unused*/, Node* node) const {
     const auto num_outputs = node->outputs().size();
-    node->i_(knum_outputs, num_outputs);
+    node->i_(knum_outputs, static_cast<int64_t>(num_outputs));
   }
 
   Node* adapt(std::shared_ptr<Graph> graph, Node* node) const override {
